+++ src/commands.rs
@@ -65,6 +65,10 @@ pub fn handle_config_command(subcommand: ConfigCommand) -> Result<(), Applicatio
     match subcommand {
         ConfigCommand::Init { force } => init_config(force).map_err(ApplicationError::from),
         ConfigCommand::Upgrade => upgrade_config().map_err(ApplicationError::ReadConfigError),
+        ConfigCommand::Show => {
+            show_config();
+            Ok(())
+        },
     }
 }
 
@@ -200,6 +204,12 @@ fn upgrade_config() -> Result<(), ReadConfigError> {
         println!("success");
     }
     result
+}
+
+/// Subcommand: config show
+fn show_config() {
+    let config = Config::from_file_or_default();
+    print!("{config}");
 }
 
 /// Used by db clean and db refresh