+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ## Unreleased
+- Added 'config show' command
 
 ## [0.2.0] - 2024-07-17
 - Added formatting options (config and args) to preserve original new lines OR use custom width